[CACHE] Move CacheTest to CoreHandlerTest
